### 2020-03-04 #2143 @whedon
[PRE REVIEW]: DetecTree: Tree detection from aerial imagery in Python
**Submitting author:** @martibosch (<a href="http://orcid.org/0000-0001-8735-9144">Martí Bosch</a>)
**Repository:** <a href="https://github.com/martibosch/detectree" target ="_blank">https://github.com/martibosch/detectree</a>
**Version:** v0.3.0
