### 2020-03-04 80e2140
[ci skip] paper corrections: tags, affiliation, date and bib dois
