+++ paper/paper.bib
@@ -5,7 +5,7 @@
   pages={23--37},
   year={1995},
   organization={Springer},
-  doi={https://doi.org/10.1007/3-540-59119-2_166}
+  doi={10.1007/3-540-59119-2_166}
 }
 
 @article{oliva2001modeling,
@@ -17,7 +17,7 @@
   pages={145--175},
   year={2001},
   publisher={Springer},
-  doi={https://doi.org/10.1023/A:1011139631724}
+  doi={10.1023/A:1011139631724}
 }
 
 @article{boykov2004experimental,
@@ -29,7 +29,7 @@
   pages={1124--1137},
   year={2004},
   publisher={IEEE},
-  doi={https://doi.org/10.1109/TPAMI.2004.60}  
+  doi={10.1109/TPAMI.2004.60}  
 }
 
 @inproceedings{yang2009tree,
@@ -38,7 +38,7 @@
   booktitle={Proceedings of the 17th ACM SIGSPATIAL International Conference on Advances in Geographic Information Systems},
   pages={131--137},
   year={2009},
-  doi={https://doi.org/10.1145/1653771.1653792}
+  doi={10.1145/1653771.1653792}
 }
 
 @inproceedings{mckinney2010data,
@@ -58,7 +58,7 @@
   number={Oct},
   pages={2825--2830},
   year={2011},
-  doi={https://doi.org/10.5555/1953048.2078195}
+  doi={10.5555/1953048.2078195}
 }
 
 @article{van2011numpy,
@@ -70,7 +70,7 @@
   pages={22},
   year={2011},
   publisher={IEEE Computer Society},
-  doi={https://doi.org/10.1109/MCSE.2011.37}
+  doi={10.1109/MCSE.2011.37}
 }
 
 @article{van2014scikit,
@@ -81,7 +81,7 @@
   pages={e453},
   year={2014},
   publisher={PeerJ Inc.},
-  doi={https://doi.org/10.7717/peerj.453}
+  doi={10.7717/peerj.453}
 }
 
 @inproceedings{rocklin2015dask,
@@ -90,7 +90,7 @@
   booktitle={Proceedings of the 14th Python in Science Conference},
   number={130-136},
   year={2015},
-  doi={https://doi.org/10.25080/Majora-7b98e3ed-013}
+  doi={10.25080/Majora-7b98e3ed-013}
 }
 
 @inproceedings{jupyter2018binder,
@@ -99,7 +99,7 @@
   booktitle={Proceedings of the 17th Python in Science Conference},
   pages={113-120},
   year={2018},
-  doi={https://doi.org/10.25080/Majora-4af1f417-011}
+  doi={10.25080/Majora-4af1f417-011}
 }
 
 @article{virtanen2020scipy,
@@ -109,5 +109,5 @@
   pages={1--12},
   year={2020},
   publisher={Nature Publishing Group},
-  doi={https://doi.org/10.1038/s41592-019-0686-2}
+  doi={10.1038/s41592-019-0686-2}
 }
