### 2020-03-04 #2143 @Kevin-Mattheus-Moerman
@terrytangyuan @arokem @kthyng could one of you edit this submission?

Some potential reviewer overlap may exists with:
https://github.com/openjournals/joss-reviews/issues/2056
https://github.com/openjournals/joss-reviews/issues/1825
