### 2020-03-04 #2143 @martibosch
the affiliation is copied from [my other JOSS paper source](https://github.com/martibosch/swisslandstats-geopy/blob/master/paper/paper.md), in which I am also the sole author. I do not know what might be wrong.
