### 2020-03-04 #2143 @arfon
> the affiliation is copied from my other JOSS paper source, in which I am also the sole author. I do not know what might be wrong.

Yes, but this isn't compliant with our updated format. Please follow the example [in our docs](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) and break the `affiliations` into their own YAML key.
