### 2020-03-04 #2143 @martibosch
Hello @arfon I am sorry about that and I thank you for your help. I have amended the paper md source and the PDF is compiled correcly now :smile:
