### 2020-03-06 #2143 @kthyng
@martibosch It looks like your documentation could be improved. I went to your readthedocs but it doesn't include the more useful documentation from your readme or from the example repo you mention.
