### 2020-03-06 #2143 @kthyng
@martibosch Could you look through the reviewer list (https://bit.ly/joss-reviewers) and give suggestions of who might be relevant reviewers for your software? Please don't use "@" to mention them so that they do not get emailed about this, just list their github username. You can also provide information for reviewers who are not on the list if you have other suggestions.
