### 2020-03-06 #2143 @martibosch
First of all thank you for editing my submission :smile:

> @martibosch It looks like your documentation could be improved. I went to your readthedocs but it doesn't include the more useful documentation from your readme or from the example repo you mention.

I understand that you mean that I should include the documentation for the technical methods described in the `background.ipynb` notebook of the example repository (e.g., the classes and methods to build the features and image descriptors)? The reason that I had not included them in the readthedocs site is because end users are not supposed to use this methods directly but through the arguments of the `TrainingSelector`, `Classifier` and `ClassifierTrainer` classes and their methods - which are documented. However it is indeed a good idea to document the other background methods as well, e.g., to help potential contributors. I will get to it tomorrow :+1:
