### 2020-03-11 c62bdca
advanced topics docs, drop `num_blocks` arg
