+++ detectree/image_descriptor.py
@@ -13,8 +13,34 @@ __all__ = [
 
 
 def compute_image_descriptor(img_rgb, kernels, response_bins_per_axis,
-                             num_blocks, num_color_bins):
+                             num_color_bins):
+    """
+    Compute a GIST descriptor for `img_rgb`. See the `background
+    <https://bit.ly/2KlCICO>`_ example notebook for more details.
+
+    Parameters
+    ----------
+    img_rgb : array-like
+        The image in RGB format, i.e., in a 3-D array
+    kernels : list-like
+        List of kernel 2-D arrays that correspond to the filter bank
+    response_bins_per_axis : int
+        Number of spatial bins per axis into which the responses to the filter
+        bank will be aggreated. For example, a value of 2 will aggregate the
+        responses into the four quadrants of the image (i.e., 2x2, 2 bins in
+        each axis of the image).
+    num_color_bins : int
+        Number of color bins per axis of the L*a*b color space with which
+        the joint color histogram will be computed
+
+    Returns
+    -------
+    img_descr : array-like
+        Vector representing GIST descriptor of `img_rgb`
+    """
+
     # gist descriptor
+    num_blocks = response_bins_per_axis**2
     gist_descr = np.zeros(len(kernels) * num_blocks)
     img_gray = color.rgb2gray(img_rgb)
     block_shape = tuple(size // response_bins_per_axis
@@ -56,8 +82,34 @@ def compute_image_descriptor(img_rgb, kernels, response_bins_per_axis,
 
 
 def compute_image_descriptor_from_filepath(img_filepath, kernels,
-                                           response_bins_per_axis, num_blocks,
+                                           response_bins_per_axis,
                                            num_color_bins):
+    """
+    Compute a GIST descriptor for `img_filepath`. See the `background
+    <https://bit.ly/2KlCICO>`_ example notebook for more details.
+
+    Parameters
+    ----------
+    img_filepath : str, file object or pathlib.Path object
+        Path to a file, URI, file object opened in binary ('rb') mode, or a
+        Path object representing the image for which a GIST descriptor will be
+        computed. The value will be passed to `rasterio.open`.
+    kernels : list-like
+        List of kernel 2-D arrays that correspond to the filter bank
+    response_bins_per_axis : int
+        Number of spatial bins per axis into which the responses to the filter
+        bank will be aggreated. For example, a value of 2 will aggregate the
+        responses into the four quadrants of the image (i.e., 2x2, 2 bins in
+        each axis of the image).
+    num_color_bins : int
+        Number of color bins per axis of the L*a*b color space with which
+        the joint color histogram will be computed
+
+    Returns
+    -------
+    img_descr : array-like
+        Vector representing GIST descriptor of `img_rgb`
+    """
     img_rgb = utils.img_rgb_from_filepath(img_filepath)
     return compute_image_descriptor(img_rgb, kernels, response_bins_per_axis,
-                                    num_blocks, num_color_bins)
+                                    num_color_bins)
