+++ docs/src/references.bib
@@ -1,8 +1,20 @@
+@article{oliva2001modeling,
+  title={Modeling the shape of the scene: A holistic representation of the spatial envelope},
+  author={Oliva, Aude and Torralba, Antonio},
+  journal={International journal of computer vision},
+  volume={42},
+  number={3},
+  pages={145--175},
+  year={2001},
+  publisher={Springer},
+  doi={10.1023/A:1011139631724}
+}
+
 @inproceedings{yang2009tree,
   title={Tree detection from aerial imagery},
   author={Yang, Lin and Wu, Xiaqing and Praun, Emil and Ma, Xiaoxu},
   booktitle={Proceedings of the 17th ACM SIGSPATIAL International Conference on Advances in Geographic Information Systems},
   pages={131--137},
   year={2009},
-  organization={ACM}
+  doi={10.1145/1653771.1653792}
 }

[…]
