+++ setup.py
@@ -5,7 +5,7 @@ from os import path
 
 from setuptools import find_packages, setup
 
-__version__ = '0.3.0'
+__version__ = '0.3.1'
 
 classifiers = [
     'License :: OSI Approved :: GNU General Public License v3 (GPLv3)',
