+++ detectree/__init__.py
@@ -2,4 +2,4 @@ from .classifier import Classifier, ClassifierTrainer
 from .train_test_split import TrainingSelector
 from .utils import split_into_tiles
 
-__version__ = '0.3.0'
+__version__ = '0.3.1'
