### 2020-03-11 #2143 @martibosch
I have added [an "Advanced Topics"](https://detectree.readthedocs.io/en/latest/advanced_topics.html) page in the documentation, which I hope addresses (or at least partly addresses) the comment of @kthyng :smiley:
