### 2020-03-19 #2143 @kthyng
> I understand that you mean that I should include the documentation for the technical methods described in the `background.ipynb` notebook of the example repository (e.g., the classes and methods to build the features and image descriptors)? The reason that I had not included them in the readthedocs site is because end users are not supposed to use this methods directly but through the arguments of the `TrainingSelector`, `Classifier` and `ClassifierTrainer` classes and their methods - which are documented. However it is indeed a good idea to document the other background methods as well, e.g., to help potential contributors. I will get to it tomorrow 👍

I mean — I am trying to learn about the package you wrote, like I have been as an editor, and I want to go somewhere to learn about it more. Normally readthedocs functions this way, but yours (previously) was strictly description of the functions. Where do you imagine a new user starting with your code? The readme seems like a good start, but also seems too short.

Anyway, this will surely come up in review too, but I figured you could start working on it :)
