### 2020-03-23 #2143 @thomakra
Hi Kristen,

Unfortunately I do not have the capacity at this time, although the project
seems interesting.

-Thomas-

On Thu, 19 Mar 2020 at 20:06, Kristen Thyng <notifications@github.com>
wrote:

> @JeffWalton-PSC <https://github.com/JeffWalton-PSC> and @thomakra
> <https://github.com/thomakra> Are you interested and willing to review
> this JOSS submission? Normally I would ask for the reviews within about 3
> weeks, but many things are unusual right now, so timing is flexible too.
> Let me know what you think. Thanks.
>
