### 2020-04-07 #2143 @kthyng
@rmsare Might you be able to review this submission to JOSS? Thanks for your consideration.
