### 2020-04-07 #2143 @kthyng
@rmsare Yes, that would be great. We are grateful for any time people can dedicate to JOSS right now in these unusual times. I will add you as the second reviewer and we can move ahead to start the review.
