### 2020-04-08 #2172 @martibosch
Regarding the DOIs: I have added the missing DOI for pandas as suggested by @whedon. I have also found that the invalid DOI (of scikit-learn) actually corresponds to a record in the digital ACM library rather than a DOI, nevertheless I could not find the DOI of the actual journal article. Shall I instead add the zenodo DOI for the library, i.e., 10.5281/zenodo.3696718?
