### 2020-04-10 0128e18
[ci skip] fix DOIs in paper references
