+++ paper/paper.bib
@@ -48,6 +48,7 @@
   volume={445},
   pages={51--56},
   year={2010},
+  doi={10.25080/majora-92bf1922-00a}
 }
 
 @article{pedregosa2011scikit,
@@ -58,7 +59,7 @@
   number={Oct},
   pages={2825--2830},
   year={2011},
-  doi={10.5555/1953048.2078195}
+  doi={10.5281/zenodo.3696718}
 }
 
 @article{van2011numpy,
