### 2020-04-10 #2172 @martibosch
I have fixed the DOIs as suggested above in [0128e18](https://github.com/martibosch/detectree/commit/0128e18f33e3380a1e209b5eac4a064f5218f783)
