### 2020-05-20 #2172 @kthyng
Hi everyone! We are officially starting JOSS back up today. We know that everyone has different circumstances with work and home life right now, so we want to figure out a way forward that will work for everyone.

@JeffWalton-PSC and @rmsare — are you both still able to review this submission? If so, what timelines would work for you?
