### 2020-05-28 martibosch/detectree#4 @rmsare
[Joss review] Add simple example
The detectree-example repository does a very nice job of walking users through the core methods implemented in detectree and several options for sampling/grouping the training data to yield useful classifiers given a variety of scenes.

It is missing an example use case, though. Something as simple as taking imagery from a particular area, training and applying the tree detection method, and summarizing the tile-level percentage of tree cover (or other useful and interesting statistics). A full end-to-end example would be best. This would give a potential user an easy entry point for their own work.

This example workflow could also be added to a "Using detectree" section of the ReadTheDocs site - a user could skim it to quickly get a sense of how the API could be used in practice.

[JOSS review](https://github.com/openjournals/joss-reviews/issues/2172)
