### 2020-05-28 martibosch/detectree#4 @rmsare
Also, you provided a Makefile for a fully reproducible workflow. What a great idea!
