### 2020-05-28 martibosch/detectree#3 @rmsare
Clicking around, I see you've produced a tree canopy map of Lausanne using detectree. This would be a nice example to cite in the paper!
