### 2020-05-28 martibosch/detectree#5 @rmsare
[Joss review] Cite related work/comparison to other methods
- [x] Add one or two sentences citing related work
- [x] Have you compared detectree to other methods? If so, it would be nice to discuss some of the pros/cons in the paper.

Related to this, the paper documents the issues surrounding lidar-based canopy height classifiers quite well.

- [x] Add citation section to decetree's README from [lausanne-tree-canopy](https://github.com/martibosch/lausanne-tree-canopy)

I like your suggested citation of both papers in the Lausanne project docs, and it would be good to add to the detectree docs.

[JOSS review](https://github.com/openjournals/joss-reviews/issues/2172)
