### 2020-05-28 #2172 @rmsare
@martibosch @kthyng Thank you for your patience with my review! I enjoyed playing around with detectree.

This is a nice package that solves a remote sensing problem - tree detection in aerial imagery. It re-implements a published method in Python and this work is appropriately referenced. The API is clear and well documented and it uses the modern Python data science stack. The functionality seems like it will be useful for spatial, social science, and environmental science research, or even commercial applications in real estate or insurance.

My main suggestions are to provide more concrete examples of use in the paper and consider reorganizing the text so readers can get a high-level overview of detectree’s research applications. Adding a simple example to the docs would lower the barrier to entry for potential users.

**Recommendation:** Accept after minor revisions

- [ ] Improve statement of need and other small edits ([#3](https://github.com/martibosch/detectree/issues/3))
- [ ] Add an example use case notebook ([#4](https://github.com/martibosch/detectree/issues/4))
- [ ] Discuss related work ([#5](https://github.com/martibosch/detectree/issues/5))

**Full disclosure:** The author reviewed a submission I made to JOSS last year, but I don't know him personally and we don't share any collaborators.
