### 2020-05-28 martibosch/detectree#3 @rmsare
I think the urban heat island example you give in your talk is a good one, too. Estimating the tree cover of road pixels in a separate dataset using detectree output would make a nice step-by-step example for the docs.
