### 2020-06-03 martibosch/detectree#3 @martibosch
I have moved the statement of need as the second paragraph and I have added a sentence mentioning a possible research problem, i.e., evaluation of the provision of ecosystem services. I have also added such statement in the README.

The methodology has now a dedicated section, which should further address #5.

Regarding the example application, the tree canopy map of Lausanne is part of our study of urban heat islands, nevertheless we do not have any proper citation for it yet - we hope to have it by the end of the month. In the meantime, I have added a sentence in the "Availability" section mentioning the Lausanne tree canopy repository.
