### 2020-06-03 bf682aa
[ci skip] paper and README update: statement of need, related work
