+++ paper/paper.bib
@@ -94,6 +94,14 @@
   doi={10.25080/Majora-7b98e3ed-013}
 }
 
+@inproceedings{he2017mask,
+  title={Mask r-cnn},
+  author={He, Kaiming and Gkioxari, Georgia and Doll{\'a}r, Piotr and Girshick, Ross},
+  booktitle={Proceedings of the IEEE international conference on computer vision},
+  pages={2961--2969},
+  year={2017}
+}
+
 @inproceedings{jupyter2018binder,
   title={Binder 2.0 - Reproducible, interactive, sharable environments for science at scale},
   author={{M}atthias {B}ussonnier and {J}essica {F}orde and {J}eremy {F}reeman and {B}rian {G}ranger and {T}im {H}ead and {C}hris {H}oldgraf and {K}yle {K}elley and {G}ladys {N}alvarte and {A}ndrew {O}sheroff and {M} {P}acer and {Y}uvi {P}anda and {F}ernando {P}erez and {B}enjamin {R}agan-{K}elley and {C}arol {W}illing},
@@ -101,6 +109,34 @@
   pages={113-120},
   year={2018},
   doi={10.25080/Majora-4af1f417-011}
+}
+
+@article{tianyang2018single,
+  title={Single-tree detection in high-resolution remote-sensing images based on a cascade neural network},
+  author={Tianyang, Dong and Jian, Zhang and Sibin, Gao and Ying, Shen and Jing, Fan},
+  journal={ISPRS International Journal of Geo-Information},
+  volume={7},
+  number={9},
+  pages={367},
+  year={2018},
+  publisher={Multidisciplinary Digital Publishing Institute}
+}
+
+@inproceedings{jain2019efficient,
+  title={An efficient framework for monitoring tree cover in an area through aerial images},
+  author={Jain, Suchit and Mittal, Rohan and Jain, Arpan and Sinha, Aakash},
+  booktitle={Applications of Machine Learning},
+  volume={11139},
+  pages={1113918},
+  year={2019},
+  organization={International Society for Optics and Photonics}
+}
+
+@misc{swisstopo2019swissimage,
+  title={SWISSIMAGE La mosa\"ique d'ortophotos de la Suisse},
+  author="{Federal Office of Topography}",
+  year=2019,
+  howpublished={Available from (in French) \url{https://shop.swisstopo.admin.ch/fr/products/images/ortho_images/SWISSIMAGE}. Accessed: 3 March 2020},
 }
 
 @article{virtanen2020scipy,
