### 2020-06-03 martibosch/detectree#5 @martibosch
I have added a paragraph about related work and a note regarding the classification accuracy comparison (which we do not have because the related works report their accuracy in terms of single-tree detection i.e., segmentation).

The README now shows the suggested citation of both papers.
