### 2020-06-03 martibosch/detectree#3 @rmsare
These changes address my concerns and have improved the paper
