### 2020-06-03 martibosch/detectree#5 @rmsare
The paper now discusses other work in detail
