### 2020-06-04 #2172 @JeffWalton-PSC
@kthyng @martibosch  Please accept my apologies for taking so long on this review. It is the first JOSS review I have been involved with. I have learned a great deal. Thank you for your patience.

DetecTree is a nice piece of software.  I ran through all of the example notebooks in the [detectree-example](https://github.com/martibosch/detectree-example) repository.  I have made some comments in the **detectree-example** issue tracker.  It took me quite a while to get all of the dependencies installed and working.  The LAS-tools required some effort.  As @rmsare has suggested, a simple end-to-end example on the main DetecTree repository would likely be helpful.  The use of LIDAR as a method to collect training data is impressive, but adds a high level of complexity for a user who is just checking out the DetecTree software.

**Recommendation:** Accept after minor edits to the paper.

**Minor edit suggestions:**
- [ ] in Step 1 "involve manual **digitizing**"
- [ ] in Step 2 "exploi**t**ed"
- [ ] "Another approach by Tianyang et al. [@tianyang2018single] employ**s**"
- [ ] "The computation ... make**s** use"

@kthyng Thank you.
