### 2020-06-10 4c34a38
[ci skip] corrected typos in paper
