### 2020-06-11 martibosch/detectree#4 @martibosch
I have added a full end-to-end example to compute a tree canopy map for the Aussersihl district in Zurich in [7533b0](https://github.com/martibosch/detectree-example/commit/7533b0db51ba6f3a4e7f1d9ac9dfdd3fb456babf). It is a small spatial extent so that the whole notebook runs in 20 minutes in my laptop.

The computation of some statistics could be appended to the notebook as well, but I find that predicting the canopy map and dumping it to a single GeoTIFF file already provides a good example of the API of DetecTree.

Both the README and ReadTheDocs site link to the example repository, which now features the new notebook. But we can add an explicit link to it if you find that it is more appropriate. In any case, as soon as we have a proper example of an academic paper using DetecTree, I will add it to the README and ReadTheDocs site as well. Actually, we have submitted a paper in which we use DetecTree to obtain a canopy map (the one in the `lausanne-tree-canopy` repository) used to evaluate urban heat islands in Lausanne, yet I do not know when we will have a proper citation format for it. I thus hope that in the meantime, the new notebook illustrates an example use case well enough.
