### 2020-06-11 #2172 @martibosch
Hello all,

first of all thank you @rmsare and @JeffWalton-PSC for your reviews.

I have addressed your issues in the `detectree` and `detectree-example` repository. An end-to-end example is now available at https://github.com/martibosch/detectree-example/blob/master/notebooks/aussersihl-canopy.ipynb On the other hand, I have corrected the typos pointed out by @JeffWalton-PSC  in [4c34a3](https://github.com/martibosch/detectree/commit/4c34a3845962985ad7eeb7fcc89177d69afc813d).

Let me know if you have any further suggestions. Best,
Martí
