### 2020-06-11 martibosch/detectree#4 @rmsare
Thanks - the new example looks good and it's easy to follow. I think adding an explicit link in the README would be useful, so users don't have to work to find it.
