### 2020-06-11 7f667a7
[ci skip] added link to Aussersihl example in README and docs index
