### 2020-06-11 martibosch/detectree#4 @martibosch
Alright then - I have added the links to the new example in the README and the documentation index in [67d7432
](https://github.com/martibosch/detectree/commit/67d74329736fa167b2c9c6fe3b9d70cb0ba2bfbc)
