### 2020-06-11 #2172 @rmsare
Thanks! These changes address my review.

I have one last minor suggestion about the README. The statement of need is a bit long and hard to read, It could be split into two sentences. So

*The target audience is researchers and practitioners in GIS that are interested in two-dimensional aspects of trees, such as their proportional abundance and spatial distribution throughout a region of study and thereafter assess important aspects of urban planning such as the provision of urban ecosystem services.*

might become something like

*The target audience is researchers and practitioners in GIS that are interested in two-dimensional aspects of trees, such as their proportional abundance and spatial distribution throughout a region of study.* **These measurements can be used to** *assess important aspects of urban planning such as the provision of urban ecosystem services.*
