### 2020-06-11 martibosch/detectree#4 @rmsare
Excellent - looks good to me
