### 2020-06-11 #2172 @martibosch
Indeed splitting the statement makes it more readable. I have amended it (both in the README and in the paper) in [003f76](https://github.com/martibosch/detectree/commit/003f736294e7eda2ff30788ea8a956e78c140c0f)
