+++ paper/paper.bib
@@ -129,7 +129,8 @@
   volume={11139},
   pages={1113918},
   year={2019},
-  organization={International Society for Optics and Photonics}
+  organization={International Society for Optics and Photonics},
+  doi={10.1117/12.2529442}
 }
 
 @misc{swisstopo2019swissimage,
