### 2020-06-24 0f4e12b
Some typo fixes
