### 2020-06-24 #2172 @kthyng
Hi @martibosch, I have some small changes in PR [#6](https://github.com/martibosch/detectree/pull/6). If you agree, please merge that. A few other items from the paper:

- in Step 4: "...class might respectively..." Should this be might? I am not clear on the meaning of the sentence here.
- "Zurich and Lausanne" are mentioned, but should this be a citation?
- Can you please do a detailed check of the references to be sure they are accurate? In particular, check the capitalization of the titles and journals. You can preserve capitalization in your bibtex entries by adding {} around the words in which to preserve it.
