### 2020-06-24 1516fa6
[ci skip] capitalization in bibtex references
