+++ paper/paper.bib
@@ -9,7 +9,7 @@
 }
 
 @article{oliva2001modeling,
-  title={Modeling the shape of the scene: A holistic representation of the spatial envelope},
+  title={Modeling the shape of the scene: {A} holistic representation of the spatial envelope},
   author={Oliva, Aude and Torralba, Antonio},
   journal={International journal of computer vision},
   volume={42},
@@ -42,7 +42,7 @@
 }
 
 @inproceedings{mckinney2010data,
-  title={Data structures for statistical computing in python},
+  title={Data structures for statistical computing in {P}ython},
   author={McKinney, Wes and others},
   booktitle={Proceedings of the 9th Python in Science Conference},
   volume={445},
@@ -52,7 +52,7 @@
 }
 
 @article{pedregosa2011scikit,
-  title={Scikit-learn: Machine learning in Python},
+  title={Scikit-learn: {M}achine learning in {P}ython},
   author={Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and others},
   journal={Journal of machine learning research},
   volume={12},
@@ -63,7 +63,7 @@
 }
 
 @article{van2011numpy,
-  title={The NumPy array: a structure for efficient numerical computation},
+  title={The {N}um{P}y array: a structure for efficient numerical computation},
   author={Van Der Walt, Stefan and Colbert, S Chris and Varoquaux, Gael},
   journal={Computing in Science \& Engineering},
   volume={13},
@@ -75,7 +75,7 @@
 }
 
 @article{van2014scikit,
-  title={scikit-image: image processing in Python},
+  title={scikit-image: image processing in {P}ython},
   author={Van der Walt, Stefan and Sch{\"o}nberger, Johannes L and Nunez-Iglesias, Juan and Boulogne, Fran{\c{c}}ois and Warner, Joshua D and Yager, Neil and Gouillart, Emmanuelle and Yu, Tony},
   journal={PeerJ},
   volume={2},
@@ -95,7 +95,7 @@
 }
 
 @inproceedings{he2017mask,
-  title={Mask r-cnn},
+  title={Mask {R}-{CNN}},
   author={He, Kaiming and Gkioxari, Georgia and Doll{\'a}r, Piotr and Girshick, Ross},
   booktitle={Proceedings of the IEEE international conference on computer vision},
   pages={2961--2969},
@@ -103,7 +103,7 @@
 }
 
 @inproceedings{jupyter2018binder,
-  title={Binder 2.0 - Reproducible, interactive, sharable environments for science at scale},
+  title={Binder 2.0 - {R}eproducible, interactive, sharable environments for science at scale},
   author={{M}atthias {B}ussonnier and {J}essica {F}orde and {J}eremy {F}reeman and {B}rian {G}ranger and {T}im {H}ead and {C}hris {H}oldgraf and {K}yle {K}elley and {G}ladys {N}alvarte and {A}ndrew {O}sheroff and {M} {P}acer and {Y}uvi {P}anda and {F}ernando {P}erez and {B}enjamin {R}agan-{K}elley and {C}arol {W}illing},
   booktitle={Proceedings of the 17th Python in Science Conference},
   pages={113-120},
@@ -134,14 +134,14 @@
 }
 
 @misc{swisstopo2019swissimage,
-  title={SWISSIMAGE La mosa\"ique d'ortophotos de la Suisse},
+  title={{SWISSIMAGE} La mosa\"ique d'ortophotos de la {S}uisse},
[…]
