### 2020-06-24 #2172 @martibosch
Hello @kthyng, I have merged your PR, thank you for your corrections. Regarding your three comments:

- I believe that I was taught that *might* and *can* are synonyms in this context, but I am no native English speaker, so we can use *can* as set in your PR.
- I suppose that you refer to the part of:
    > New York achieve a pixel classification accuracy of 91.7%, whereas the example applications of DetecTree in Zurich and Lausanne achieve accuracies of 85.98% and 91.75% respectively

  In such case, the accuracy scores are taken from GitHub repositories (i.e., [detectree-example](https://github.com/martibosch/detectree-example) and [lausanne-tree-canopy](https://github.com/martibosch/lausanne-tree-canopy) for which I do not have any proper citation. I hope to have a citation for the Lausanne tree canopy soon, since such a canopy map was produced in the context of a study of the urban heat islands which has been submitted to [GMD](https://www.geoscientific-model-development.net). Therefore, should I leave it like that, with no citation? Alternatively, I can archive both repositories to Zenodo so that we can cite them in this manuscript.
- All references should be properly formatted as of [1516fa6](https://github.com/martibosch/detectree/commit/1516fa694b8ef868f23632ca2c8663ad6bc0bf54)
