### 2020-06-25 #2172 @kthyng
Hi @martibosch!

- "might" has a bit of uncertainty to it usually, so "can" is better here I think. Is this change in there?
- Ah, I see my misunderstanding. I read it as two people's last names instead of two places! Sorry about that. I haven't been to Switzerland before, sadly. You can choose what makes the most sense here.

Please let me know when this is all accounted for, thanks.
