### 2020-06-25 #2172 @martibosch
Hello @kthyng ,

- the "might" was changed to "can" in your PR, so since I merged it, the current version reads
  > Furthermore, the functionalities of DetecTree *can* be used through its Python API as well as through its command-line interface (CLI)
- in the absence of a proper academic article to cite, I believe that we can leave the sentence as it is.

I hope that this addresses the two remarks.
