### 2020-06-25 #2172 @kthyng
I made a change in my PR for one "might" but there was another under Step 4 that I hadn't been sure about at the time. Can you change that one?

Next, can you verify the version, and then submit your code to an archive like Zenodo and report the doi here? Make sure that the title and author list on the Zenodo archive match those of your JOSS paper (you may have to alter the Zenodo metadata to do this).
