### 2020-06-25 3ab0571
[ci skip] changed "might" for "can" in Step 4 of the paper
