### 2020-06-25 #2172 @martibosch
I have checked for all the occurences of "might" in the manuscript and replaced the one at Step 4 in [3ab057](https://github.com/martibosch/detectree/commit/3ab057148506fb07701c18b4564c7443a78267ba).

I have created a released named `joss-paper` for the version v0.3.1 of DetecTree and archived it in Zenodo at [10.5281/zenodo.3908338](https://doi.org/10.5281/zenodo.3908338). Finally, I edited the title and author list from the Zenodo archive so that they now match those of the JOSS paper.
